Return exact lastModified date from response headers
Apache doesn’t compare if the date is after the modification date, it
juste compares if the dates are equal, so it fails when the date is in
the future of the Last-Modified header sent in the response.

Rename the option to match the name of the header (lastCheckedAt ->
lastModified).